feat(turndown): add WASM binding via core split
Extracts the full html5ever-based HTML → Markdown converter (parser
walker, list / table / code-block handling, GFM rules) into
`_turndown-core`. The napi crate is now a thin shim wrapping
`core::turndown(&str, &TurndownOptions)`; WASM exposes the same
surface plus `turndownBatch` via wasm-bindgen + serde-wasm-bindgen.
