+++ crates/turndown/Cargo.toml
@@ -9,9 +9,7 @@ crate-type = ["cdylib"]
 [dependencies]
 napi = { workspace = true }
 napi-derive = { workspace = true }
-html5ever = "0.39"
-markup5ever = "0.39"
-markup5ever_rcdom = "0.39"
+amigo-turndown-core = { path = "../_turndown-core" }
 
 [build-dependencies]
 napi-build = "2"
