+++ crates/turndown/wasm/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "amigo-turndown-wasm"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+amigo-turndown-core = { path = "../../_turndown-core" }
+wasm-bindgen = { workspace = true }
+serde-wasm-bindgen = { workspace = true }
+serde = { version = "1", features = ["derive"] }
+
+[dev-dependencies]
+wasm-bindgen-test = { workspace = true }
+
+[package.metadata.wasm-pack.profile.release]
+wasm-opt = false
